Translate DBOMembershipRequests in migration
Rename DBOMembershipInvtnSubmission to DBOMembershipInvitation.
Rename DBOMembershipRqstSubmission to DBOMembershipRequest.
The table names were left unchanged for fear of breaking migration.
Refactor zip and unzip util methods into their own ZipUtils class.
