+++ lib/jdomodels/src/main/java/org/sagebionetworks/repo/model/dbo/dao/MembershipInvitationDAOImpl.java
@@ -5,19 +5,18 @@ package org.sagebionetworks.repo.model.dbo.dao;
 
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_GROUP_MEMBERS_GROUP_ID;
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_GROUP_MEMBERS_MEMBER_ID;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_CREATED_ON;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_ETAG;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_ID;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_EMAIL;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_PROPERTIES;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_CREATED_ON;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_ETAG;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_EXPIRES_ON;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_ID;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_INVITEE_ID;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_PROPERTIES;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_MEMBERSHIP_INVITATION_TEAM_ID;
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.COL_TEAM_ID;
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.LIMIT_PARAM_NAME;
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.OFFSET_PARAM_NAME;
 import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.TABLE_GROUP_MEMBERS;
-import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.TABLE_MEMBERSHIP_INVITATION_SUBMISSION;
+import static org.sagebionetworks.repo.model.query.jdo.SqlConstants.TABLE_MEMBERSHIP_INVITATION;
 
 import java.sql.Blob;
 import java.util.List;
@@ -31,7 +30,7 @@ import org.sagebionetworks.repo.model.InvalidModelException;
 import org.sagebionetworks.repo.model.MembershipInvitation;
 import org.sagebionetworks.repo.model.MembershipInvitationDAO;
 import org.sagebionetworks.repo.model.dbo.DBOBasicDao;
-import org.sagebionetworks.repo.model.dbo.persistence.DBOMembershipInvtnSubmission;
+import org.sagebionetworks.repo.model.dbo.persistence.DBOMembershipInvitation;
 import org.sagebionetworks.repo.transactions.WriteTransactionReadCommitted;
 import org.sagebionetworks.repo.web.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -54,38 +53,36 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	private NamedParameterJdbcTemplate namedJdbcTemplate;
 	@Autowired
 	GroupMembersDAO groupMembersDAO;
-	
-	private static final String INVITEE_ID_COLUMN_LABEL = "INVITEE_ID";
 
-	private static final String SELECT_OPEN_INVITATIONS_CORE = 
-			" FROM "+TABLE_MEMBERSHIP_INVITATION_SUBMISSION+" mis "
-			+" WHERE mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID+" NOT IN ("
+	private static final String SELECT_OPEN_INVITATIONS_CORE =
+			" FROM "+ TABLE_MEMBERSHIP_INVITATION +" mis "
+			+" WHERE mis."+ COL_MEMBERSHIP_INVITATION_TEAM_ID +" NOT IN ("
 					+ "SELECT "+COL_GROUP_MEMBERS_GROUP_ID
 					+" FROM "+TABLE_GROUP_MEMBERS
-					+" WHERE "+COL_GROUP_MEMBERS_MEMBER_ID+"=mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID
+					+" WHERE "+COL_GROUP_MEMBERS_MEMBER_ID+"=mis."+ COL_MEMBERSHIP_INVITATION_INVITEE_ID
 					+" ) "
-			+" AND ( mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON+" IS NULL"
-					+ " OR mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON+">:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON+" ) ";
+			+" AND ( mis."+ COL_MEMBERSHIP_INVITATION_EXPIRES_ON +" IS NULL"
+					+ " OR mis."+ COL_MEMBERSHIP_INVITATION_EXPIRES_ON +">:"+ COL_MEMBERSHIP_INVITATION_EXPIRES_ON +" ) ";
 	private static final String SELECT_OPEN_INVITATIONS_BY_USER_CORE =
 			SELECT_OPEN_INVITATIONS_CORE
-			+" AND mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID+"=:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID;
+			+" AND mis."+ COL_MEMBERSHIP_INVITATION_INVITEE_ID +"=:"+ COL_MEMBERSHIP_INVITATION_INVITEE_ID;
 	
 	private static final String SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_CORE = 
 			SELECT_OPEN_INVITATIONS_BY_USER_CORE
-			+" AND mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID+"=:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID;
+			+" AND mis."+ COL_MEMBERSHIP_INVITATION_TEAM_ID +"=:"+ COL_MEMBERSHIP_INVITATION_TEAM_ID;
 
 	private static final String SELECT_OPEN_INVITATIONS_BY_TEAM_CORE = 
 			SELECT_OPEN_INVITATIONS_CORE
-			+" AND mis."+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID+"=:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID;
+			+" AND mis."+ COL_MEMBERSHIP_INVITATION_TEAM_ID +"=:"+ COL_MEMBERSHIP_INVITATION_TEAM_ID;
 
 	private static final String SELECT_OPEN_INVITATIONS_BY_USER_PAGINATED = 
 			"SELECT mis.* "+SELECT_OPEN_INVITATIONS_BY_USER_CORE
-			+" ORDER BY "+COL_MEMBERSHIP_INVITATION_SUBMISSION_CREATED_ON+" DESC "
+			+" ORDER BY "+ COL_MEMBERSHIP_INVITATION_CREATED_ON +" DESC "
 			+" LIMIT :"+LIMIT_PARAM_NAME+" OFFSET :"+OFFSET_PARAM_NAME;
 	
 	private static final String SELECT_OPEN_INVITATIONS_BY_TEAM_PAGINATED = 
 			"SELECT mis.* "+SELECT_OPEN_INVITATIONS_BY_TEAM_CORE
-			+" ORDER BY "+COL_MEMBERSHIP_INVITATION_SUBMISSION_CREATED_ON+" DESC "
+			+" ORDER BY "+ COL_MEMBERSHIP_INVITATION_CREATED_ON +" DESC "
 			+" LIMIT :"+LIMIT_PARAM_NAME+" OFFSET :"+OFFSET_PARAM_NAME;
 	
 	private static final String SELECT_OPEN_INVITATIONS_BY_USER_COUNT = 
@@ -94,11 +91,11 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	private static final String SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_PAGINATED = 
 			"SELECT mis.* "
 			+SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_CORE
-			+" ORDER BY "+COL_MEMBERSHIP_INVITATION_SUBMISSION_CREATED_ON+" DESC "
+			+" ORDER BY "+ COL_MEMBERSHIP_INVITATION_CREATED_ON +" DESC "
 			+" LIMIT :"+LIMIT_PARAM_NAME+" OFFSET :"+OFFSET_PARAM_NAME;
 	
 	private static final String SELECT_INVITER_BY_TEAM_AND_USER = 
-			"SELECT "+COL_MEMBERSHIP_INVITATION_SUBMISSION_PROPERTIES+
+			"SELECT "+ COL_MEMBERSHIP_INVITATION_PROPERTIES +
 			SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_CORE;
 	
 	private static final String SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_COUNT = 
@@ -108,20 +105,15 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 			"SELECT COUNT(*) "+SELECT_OPEN_INVITATIONS_BY_TEAM_CORE;
 	
 	private static final String DELETE_INVITATIONS_BY_TEAM_AND_INVITEE = 
-			"DELETE FROM "+TABLE_MEMBERSHIP_INVITATION_SUBMISSION
-			+" WHERE "+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID+"=:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID
-			+" AND "+COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID+"=:"+COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID;
-
-	private static final String SELECT_INVITEE_EMAIL =
-			"SELECT " + COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_EMAIL + " " +
-			"FROM " + TABLE_MEMBERSHIP_INVITATION_SUBMISSION + " " +
-			"WHERE " + COL_MEMBERSHIP_INVITATION_SUBMISSION_ID + " = :" + COL_MEMBERSHIP_INVITATION_SUBMISSION_ID;
+			"DELETE FROM "+ TABLE_MEMBERSHIP_INVITATION
+			+" WHERE "+ COL_MEMBERSHIP_INVITATION_TEAM_ID +"=:"+ COL_MEMBERSHIP_INVITATION_TEAM_ID
+			+" AND "+ COL_MEMBERSHIP_INVITATION_INVITEE_ID +"=:"+ COL_MEMBERSHIP_INVITATION_INVITEE_ID;
 
 	private static final String UPDATE_INVITEE_ID =
-			"UPDATE " + TABLE_MEMBERSHIP_INVITATION_SUBMISSION + " " +
-			"SET " + COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID + " = :" + COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID + ", " +
-			COL_MEMBERSHIP_INVITATION_SUBMISSION_ETAG + " = :" + COL_MEMBERSHIP_INVITATION_SUBMISSION_ETAG + " " +
-			"WHERE " + COL_MEMBERSHIP_INVITATION_SUBMISSION_ID + " = :" + COL_MEMBERSHIP_INVITATION_SUBMISSION_ID;
+			"UPDATE " + TABLE_MEMBERSHIP_INVITATION + " " +
+			"SET " + COL_MEMBERSHIP_INVITATION_INVITEE_ID + " = :" + COL_MEMBERSHIP_INVITATION_INVITEE_ID + ", " +
+					COL_MEMBERSHIP_INVITATION_ETAG + " = :" + COL_MEMBERSHIP_INVITATION_ETAG + " " +
+			"WHERE " + COL_MEMBERSHIP_INVITATION_ID + " = :" + COL_MEMBERSHIP_INVITATION_ID;
 
 	/* (non-Javadoc)
 	 * @see org.sagebionetworks.repo.model.MemberRqstSubmissionDAO#create(org.sagebionetworks.repo.model.MemberRqstSubmission)
@@ -131,7 +123,7 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public MembershipInvitation create(MembershipInvitation dto) throws DatastoreException,
 	InvalidModelException {
 		if (dto.getId()==null) dto.setId(idGenerator.generateNewId(IdType.MEMBERSHIP_INVITATION_ID).toString());
-		DBOMembershipInvtnSubmission dbo = new DBOMembershipInvtnSubmission();
+		DBOMembershipInvitation dbo = new DBOMembershipInvitation();
 		MembershipInvitationUtils.copyDtoToDbo(dto, dbo);
 		dbo.setEtag(UUID.randomUUID().toString());
 		dbo = basicDao.createNew(dbo);
@@ -147,7 +139,7 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public MembershipInvitation get(String id) throws DatastoreException, NotFoundException {
 		MapSqlParameterSource param = new MapSqlParameterSource();
 		param.addValue(COL_TEAM_ID.toLowerCase(), id);
-		DBOMembershipInvtnSubmission dbo = basicDao.getObjectByPrimaryKey(DBOMembershipInvtnSubmission.class, param);
+		DBOMembershipInvitation dbo = basicDao.getObjectByPrimaryKey(DBOMembershipInvitation.class, param);
 		MembershipInvitation dto = MembershipInvitationUtils.copyDboToDto(dbo);
 		return dto;
 	}
@@ -156,7 +148,7 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public MembershipInvitation getWithUpdateLock(String id) {
 		MapSqlParameterSource param = new MapSqlParameterSource();
 		param.addValue(COL_TEAM_ID.toLowerCase(), id);
-		DBOMembershipInvtnSubmission dbo = basicDao.getObjectByPrimaryKeyWithUpdateLock(DBOMembershipInvtnSubmission.class, param);
+		DBOMembershipInvitation dbo = basicDao.getObjectByPrimaryKeyWithUpdateLock(DBOMembershipInvitation.class, param);
 		MembershipInvitation dto = MembershipInvitationUtils.copyDboToDto(dbo);
 		return dto;
 	}
@@ -168,20 +160,20 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public void delete(String id) throws DatastoreException, NotFoundException {
 		MapSqlParameterSource param = new MapSqlParameterSource();
 		param.addValue(COL_TEAM_ID.toLowerCase(), id);
-		basicDao.deleteObjectByPrimaryKey(DBOMembershipInvtnSubmission.class, param);
+		basicDao.deleteObjectByPrimaryKey(DBOMembershipInvitation.class, param);
 	}
 	
 	private static final RowMapper<String> INVITER_ROW_MAPPER = (rs, rowNum) -> {
-		Blob misProperties = rs.getBlob(COL_MEMBERSHIP_INVITATION_SUBMISSION_PROPERTIES);
+		Blob misProperties = rs.getBlob(COL_MEMBERSHIP_INVITATION_PROPERTIES);
 		MembershipInvitation mis = MembershipInvitationUtils.deserialize(misProperties.getBytes(1, (int) misProperties.length()));
 		return mis.getCreatedBy();
 	};
 	
-	private static final RowMapper<DBOMembershipInvtnSubmission> dboMembershipInvtnSubmissionRowMapper =
-			(new DBOMembershipInvtnSubmission()).getTableMapping();
+	private static final RowMapper<DBOMembershipInvitation> dboMembershipInvtnSubmissionRowMapper =
+			(new DBOMembershipInvitation()).getTableMapping();
 		
 	private static final RowMapper<MembershipInvitation> membershipInvitationRowMapper = (rs, rowNum) -> {
-		DBOMembershipInvtnSubmission dbo = dboMembershipInvtnSubmissionRowMapper.mapRow(rs, rowNum);
+		DBOMembershipInvitation dbo = dboMembershipInvtnSubmissionRowMapper.mapRow(rs, rowNum);
 		return MembershipInvitationUtils.copyDboToDto(dbo);
 	};
 	
@@ -189,11 +181,11 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public List<MembershipInvitation> getOpenByUserInRange(long userId, long now, long limit,
 	                                                       long offset) throws DatastoreException {
 		MapSqlParameterSource param = new MapSqlParameterSource();
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, userId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, userId);
 		param.addValue(OFFSET_PARAM_NAME, offset);
 		if (limit<=0) throw new IllegalArgumentException("'to' param must be greater than 'from' param.");
 		param.addValue(LIMIT_PARAM_NAME, limit);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.query(SELECT_OPEN_INVITATIONS_BY_USER_PAGINATED, param, membershipInvitationRowMapper);
 	}
 
@@ -201,11 +193,11 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public List<MembershipInvitation> getOpenByTeamInRange(
 			long teamId, long now, long limit, long offset) {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
 		param.addValue(OFFSET_PARAM_NAME, offset);
 		if (limit<=0) throw new IllegalArgumentException("'to' param must be greater than 'from' param.");
 		param.addValue(LIMIT_PARAM_NAME, limit);	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.query(SELECT_OPEN_INVITATIONS_BY_TEAM_PAGINATED, param, membershipInvitationRowMapper);
 	}
 
@@ -214,21 +206,21 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 			long teamId, long userId, long now, long limit,
 			long offset, RowMapper<T> rowMapper) {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, userId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, userId);
 		param.addValue(OFFSET_PARAM_NAME, offset);
 		if (limit<=0) throw new IllegalArgumentException("'to' param must be greater than 'from' param.");
 		param.addValue(LIMIT_PARAM_NAME, limit);	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);	
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.query(SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_PAGINATED, param, rowMapper);
 	}
 	
 	@Override
 	public List<String> getInvitersByTeamAndUser(long teamId, long userId, long now) {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, userId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);	
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, userId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.query(SELECT_INVITER_BY_TEAM_AND_USER, param, INVITER_ROW_MAPPER);
 	}
 
@@ -244,8 +236,8 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public long getOpenByUserCount(long userId, long now)
 			throws DatastoreException, NotFoundException {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, userId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, userId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.queryForObject(SELECT_OPEN_INVITATIONS_BY_USER_COUNT, param, Long.class);
 	}
 
@@ -254,9 +246,9 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public long getOpenByTeamAndUserCount(long teamId, long userId,
 			long now) throws DatastoreException, NotFoundException {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, userId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);	
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, userId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.queryForObject(SELECT_OPEN_INVITATIONS_BY_TEAM_AND_USER_COUNT, param, Long.class);
 	}
 
@@ -265,8 +257,8 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public long getOpenByTeamCount(long teamId,
 			long now) throws DatastoreException, NotFoundException {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON, now);	
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_EXPIRES_ON, now);
 		return namedJdbcTemplate.queryForObject(SELECT_OPEN_INVITATIONS_BY_TEAM_COUNT, param, Long.class);
 	}
 
@@ -275,8 +267,8 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	public void deleteByTeamAndUser(long teamId, long inviteeId)
 			throws DatastoreException {
 		MapSqlParameterSource param = new MapSqlParameterSource();	
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID, teamId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, inviteeId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_TEAM_ID, teamId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, inviteeId);
 		namedJdbcTemplate.update(DELETE_INVITATIONS_BY_TEAM_AND_INVITEE, param);
 	}
 
@@ -284,9 +276,9 @@ public class MembershipInvitationDAOImpl implements MembershipInvitationDAO {
 	@Override
 	public void updateInviteeId(String misId, long inviteeId) {
 		MapSqlParameterSource param = new MapSqlParameterSource();
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID, inviteeId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_ID, misId);
-		param.addValue(COL_MEMBERSHIP_INVITATION_SUBMISSION_ETAG, UUID.randomUUID().toString());
+		param.addValue(COL_MEMBERSHIP_INVITATION_INVITEE_ID, inviteeId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_ID, misId);
+		param.addValue(COL_MEMBERSHIP_INVITATION_ETAG, UUID.randomUUID().toString());
 		int rowsUpdated = namedJdbcTemplate.update(UPDATE_INVITEE_ID, param);
 		if (rowsUpdated == 0) {
 			throw new NotFoundException("Could not update MembershipInvitation with id " + misId + ".");
