+++ lib/jdomodels/src/main/java/org/sagebionetworks/repo/model/query/jdo/SqlConstants.java
@@ -658,26 +658,26 @@ public class SqlConstants {
 	public static final String COL_NOTIFICATION_EMAIL_ALIAS_ID		= "ALIAS_ID";
 	
 	// MembershipInvitation Table
-	public static final String TABLE_MEMBERSHIP_INVITATION_SUBMISSION	= "MEMBERSHIP_INVITATION_SUBMISSION";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_ID				= "ID";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_ETAG			= "ETAG";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_CREATED_ON		= "CREATED_ON";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_TEAM_ID			= "TEAM_ID";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_EXPIRES_ON		= "EXPIRES_ON";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_ID		= "INVITEE_ID";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_INVITEE_EMAIL	= "INVITEE_EMAIL";
-	public static final String COL_MEMBERSHIP_INVITATION_SUBMISSION_PROPERTIES		= "PROPERTIES";
-	public static final String DDL_FILE_MEMBERSHIP_INVITATION_SUBMISSION	= "schema/MembershipInvitationSubmission-ddl.sql";
+	public static final String TABLE_MEMBERSHIP_INVITATION = "MEMBERSHIP_INVITATION_SUBMISSION";
+	public static final String COL_MEMBERSHIP_INVITATION_ID = "ID";
+	public static final String COL_MEMBERSHIP_INVITATION_ETAG = "ETAG";
+	public static final String COL_MEMBERSHIP_INVITATION_CREATED_ON = "CREATED_ON";
+	public static final String COL_MEMBERSHIP_INVITATION_TEAM_ID = "TEAM_ID";
+	public static final String COL_MEMBERSHIP_INVITATION_EXPIRES_ON = "EXPIRES_ON";
+	public static final String COL_MEMBERSHIP_INVITATION_INVITEE_ID = "INVITEE_ID";
+	public static final String COL_MEMBERSHIP_INVITATION_INVITEE_EMAIL = "INVITEE_EMAIL";
+	public static final String COL_MEMBERSHIP_INVITATION_PROPERTIES = "PROPERTIES";
+	public static final String DDL_FILE_MEMBERSHIP_INVITATION = "schema/MembershipInvitation-ddl.sql";
 	
 	// MembershipRequest Table
-	public static final String TABLE_MEMBERSHIP_REQUEST_SUBMISSION	= "MEMBERSHIP_REQUEST_SUBMISSION";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_ID				= "ID";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_CREATED_ON		= "CREATED_ON";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_TEAM_ID		= "TEAM_ID";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_USER_ID		= "USER_ID";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_EXPIRES_ON		= "EXPIRES_ON";
-	public static final String COL_MEMBERSHIP_REQUEST_SUBMISSION_PROPERTIES		= "PROPERTIES";
-	public static final String DDL_FILE_MEMBERSHIP_REQUEST_SUBMISSION	= "schema/MembershipRequestSubmission-ddl.sql";
+	public static final String TABLE_MEMBERSHIP_REQUEST = "MEMBERSHIP_REQUEST_SUBMISSION";
+	public static final String COL_MEMBERSHIP_REQUEST_ID = "ID";
+	public static final String COL_MEMBERSHIP_REQUEST_CREATED_ON = "CREATED_ON";
+	public static final String COL_MEMBERSHIP_REQUEST_TEAM_ID = "TEAM_ID";
+	public static final String COL_MEMBERSHIP_REQUEST_USER_ID = "USER_ID";
+	public static final String COL_MEMBERSHIP_REQUEST_EXPIRES_ON = "EXPIRES_ON";
+	public static final String COL_MEMBERSHIP_REQUEST_PROPERTIES = "PROPERTIES";
+	public static final String DDL_FILE_MEMBERSHIP_REQUEST = "schema/MembershipRequest-ddl.sql";
 
 	// QuizResponse data set
 	public static final String TABLE_QUIZ_RESPONSE = "QUIZ_RESPONSE";
