+++ lib/lib-auto-generated/src/main/resources/schema/org/sagebionetworks/repo/model/migration/MigrationType.json
@@ -30,8 +30,8 @@
         "DOCKER_REPOSITORY_NAME",
         "DOCKER_COMMIT",
         "TEAM",
-        "MEMBERSHIP_INVITATION_SUBMISSION",
-        "MEMBERSHIP_REQUEST_SUBMISSION",
+        "MEMBERSHIP_INVITATION",
+        "MEMBERSHIP_REQUEST",
         "EVALUATION",
         "EVALUATION_SUBMISSIONS",
         "SUBMISSION",
